@@ -1,12 +1,12 @@
-const { Schema, model, Types } = require('mongoose');
+import { Schema, model, Types } from 'mongoose';
 
 const schema = new Schema({
-    from: {type: String, required: true},
-    to: {type: String, required: true, unique: true},
-    code: {type: String, required: true, unique: true},
-    date: {type: Date, default: Date.now},
-    clicks: {type: Number, default: 0},
-    owner: {type: Types.ObjectId, ref: 'User'}
-})
+    from: { type: String, required: true },
+    to: { type: String, required: true, unique: true },
+    code: { type: String, required: true, unique: true },
+    date: { type: Date, default: Date.now },
+    clicks: { type: Number, default: 0 },
+    owner: { type: Types.ObjectId, ref: 'User' }
+});
 
-module.exports = model('Link', schema);
+export default model('Link', schema);
