@@ -1,20 +1,20 @@
-import React, { useEffect, useRef } from 'react';
+import React, {useEffect, useRef} from 'react';
 import { useLocation } from 'react-router-dom';
 
-const HeaderWrapper = ({children}) => {
-    const headerRef = useRef(null);
-    const location = useLocation();
+const HeaderWrapper = ({ children }) => {
+  const headerRef = useRef(null);
+  const location = useLocation();
 
-    useEffect(() => {
-        console.log(`You changed the page to: ${location.pathname}`);
-        if (headerRef.current.contains(document.activeElement)) {
-            document.activeElement.blur()
-        }
-    }, [location]);
+  useEffect(() => {
+    console.log(`You changed the page to: ${location.pathname}`);
+    if (headerRef.current.contains(document.activeElement)) {
+      document.activeElement.blur()
+    }
+  }, [location]);
 
   return (
     <header tabIndex="0" ref={headerRef}>
-        {children}
+      {children}
     </header>
   );
 };
