@@ -1,32 +1,34 @@
-import React, {useRef, useState } from 'react';
+import React, { useContext, useRef } from 'react';
 import { useNavigate } from 'react-router-dom';
-// import AuthContext from '../context/AuthContext';
-import { useHttp } from '../hooks/http.hook';
 import LoginForm from '../components/LoginForm';
+import { useHttp } from '../hooks/http.hook';
+import AuthContext from '../context/AuthContext';
 
 const LoginPage = () => {
-    // const { request, error } = useHttp();
-    const formRef = useRef(null);
-    const navigate = useNavigate();
-    // const auth = useContext(AuthContext);
+  const { request, error } = useHttp();
+  const formRef = useRef(null);
+  const navigate = useNavigate();
+  const auth = useContext(AuthContext);
 
-    const loginHandler = async (formData) => {
-        const user = Object.fromEntries(formData);
-        const data = await request('/api/auth/login', 'POST', {...user});
+  const loginHandler = async (e) => {
+    e.preventDefault()
+    const formData = new FormData(formRef.current)
+    const user = Object.fromEntries(formData);
+    const data = await request('/api/auth/login', 'POST', {...user});
 
-        formRef.current.reset()
-        if (!data?.token) return;
+    formRef.current.reset()
+    if (!data?.token) return;
 
-        auth.login(data.token);
-        navigate('/recipes')
+    auth.login(data.token);
+    navigate('/recipes')
   }
 
   return (
-      <LoginForm
-          loginHandler={loginHandler}
-          // error={error}
-          ref={formRef}
-      />
+    <LoginForm
+      loginHandler={loginHandler}
+      // error={error}
+      ref={formRef}
+    />
   )
 
 };
