@@ -4,30 +4,28 @@ import SignUpForm from '../components/SignUpForm';
 
 const SignUpPage = () => {
   // const { request, error } = useHttp();
-    const formRef = useRef(null);
-    const signUpHandler = async (formData) => {
-      console.log(...formData.entries())
-  const user = Object.fromEntries(formData);
-        //console.log(...user)
-  // const data = await request('/api/auth/register', 'POST', user);
-        console.log(data)
+  const formRef = useRef(null);
+  const signUpHandler = async (formData) => {
+    console.log(...formData.entries())
+    const user = Object.fromEntries(formData);
+    // const data = await request('/api/auth/register', 'POST', user);
+    console.log(data)
 
-        // formRef.current.reset()
-        //if (!data?.token) return;
+    // formRef.current.reset()
+    //if (!data?.token) return;
 
-        //auth.login(data.token);
-        //navigate('/recipes')
+    //auth.login(data.token);
+    //navigate('/recipes')
   }
 
   const confirmHandler = (e) => {
-        console.log('blured!!!!')
-      console.log(formRef.current)
+    console.log(formRef.current)
   }
 
   return (
     <div className="form-container">
       <div>Sign Up</div>
-      <SignUpForm signUpHandler={signUpHandler} confirmHandler={confirmHandler} formRef={formRef} />
+      <SignUpForm signUpHandler={signUpHandler} confirmHandler={confirmHandler} formRef={formRef}/>
     </div>
   )
 }
